// src/app/core/models/catalogs/bank-movements-types.models.ts
export interface IBankMovementsTypes {
  id: number;
  accountKey: number;
  accountType: string;
  coinKey: string;
  square: string;
  branchOffice: number;
  rateCalculationInterest: string;
  bankKey: string;
  accountNumberTransfer: string;
  registryNumber: number;
  delegationNumber: number;
  esReference: string;
}